+++ src/pages/CartPage/CartPage.js
@@ -5,7 +5,8 @@ import GlobalStateContext from '../../contexts/GlobalStateContext';
 
 const CartPage = () => {
     let sum = 0;
-    const [cart, setCart] = useContext(GlobalStateContext)
+    const [cart,] = useContext(GlobalStateContext)
+
     const [restaurant, setRestaurant] = useState({
         "id": "1",
         "description": "Habib's é uma rede de restaurantes de comida rápida brasileira especializada em culinária árabe, os restaurantes vendem mais de 600 milhões de esfirras por ano. A empresa emprega 22 mil colaboradores e tem 421 unidades distribuídas em mais de cem municípios em 20 unidades federativas.",
@@ -16,29 +17,6 @@ const CartPage = () => {
         "deliveryTime": 60,
         "category": "Árabe"
     })
-
-    const [products, setProducts] = useState([
-        {
-            id: "CnKdjU6CyKakQDGHzNln",
-            category: "Salgado",
-            price: "1",
-            photoUrl: "https://static-images.ifood.com.br/image/upload/f_auto,t_high/pratos/65c38aa8-b094-413d-9a80-ddc256bfcc78/201907031404_66194495.jpg",
-            name: "Bibsfiha carne",
-            description: "Esfiha deliciosa, receita secreta do Habibs.",
-            amount: 1
-
-        },
-        {
-            id: "KJqMl2DxeShkSBevKVre",
-            photoUrl: "https://www.sushimanscwb.com.br/wp-content/uploads/2018/10/1579_REFRIGERANTE_LATA_-_350ml_17d2e336feb44a2696fd6cf852c41b50-1.jpeg",
-            name: "Refrigerante",
-            description: "Coca cola, Sprite ou Guaraná",
-            category: "Bebida",
-            price: "4",
-            amount: 5
-        }
-    ])
-
 
     const [profileUser, setProfileUser] = useState(
         {
@@ -52,35 +30,24 @@ const CartPage = () => {
             }
         })
 
-    const removeItem = (itemToRemove) => {
-        const position = cart.findIndex((item) => {
-            return item.id === itemToRemove.id;
-        });
 
-        let newCart = [...cart];
+    const changeAccent = (number) => {
+        return number.toFixed(2).replace(".", ",")
+    }
+    
 
-        if (newCart[position].amount === 1) {
-            newCart.splice(position, 1);
-        } else {
-            newCart[position].amount -= 1;
-        }
-
-        setCart(newCart);
-    };
-
-    const CardProduct = products.map((product) => {
+    const CardProduct = cart.map((product) => {
         sum += Number(product.price.replace(",", ".")) * product.amount
 
         const price = (Number(product.price.replace(",", ".")) * product.amount)
 
         return (
-            <ProductCard photo={product.photoUrl} name={product.name} description={product.description} price={price.toFixed(2).replace(".",",")} amount={product.amount} removeItem={removeItem}/>
+            <ProductCard photo={product.photoUrl} id={product.id} name={product.name} description={product.description} price={changeAccent(price)} amount={product.amount} />
         )
     })
 
     return (
         <ContainerPai>
-            {/* <Cabecalho /> */}
 
             <AddressContainer>
                 <AddressUser>Endereço de entrega</AddressUser>
@@ -88,23 +55,25 @@ const CartPage = () => {
             </AddressContainer>
 
             <Cart>
+                {cart.length < 1 ? <div> Carrinho vazio </div> : <>
+                    <InfoRestaurant>
+                        <span> {restaurant.name} </span>
 
-                <InfoRestaurant>
-                    <span> {restaurant.name} </span>
+                        <p>{restaurant.address}</p>
+                        <p> {restaurant.deliveryTime} - {restaurant.deliveryTime + 10} min </p>
+                    </InfoRestaurant>
 
-                    <p>{restaurant.address}</p>
-                    <p> {restaurant.deliveryTime} - {restaurant.deliveryTime + 10} min </p>
-                </InfoRestaurant>
+                    <ContainerProducts>
+                        {CardProduct}
+                    </ContainerProducts>
+                </>}
 
-                <ContainerProducts>
-                    {CardProduct}
-                </ContainerProducts>
 
                 <Info>
-                    <div>Frete R${restaurant.shipping.toFixed(2).replace(".", ",")}</div>
+                    <div>Frete R${cart.length < 1 ? <> 0,00 </> : changeAccent(restaurant.shipping)}</div>
                     <Price>
                         <p>SUBTOTAL</p>
-                        <span>R${(sum + restaurant.shipping).toFixed(2)}</span>
+                        <span>R${cart.length < 1 ? <> 0,00 </> : (changeAccent(sum + restaurant.shipping))}</span>
                     </Price>
                 </Info>
                 <Payment>
